requirements.txt
pyautogen
fastapi
uvicorn[standard]
requests
